Fix tiles being out of place when rotated
Add IsoRotation::getSpriteToRender, for subclasses to make small
additions without copying hte entire method. Use this in
IsoBuildingRenderer to draw the sprite at the center of the tile. Set
the origin of all tile and building sprites to the center of the tile in
order to still be positioned correctly when rotated.